ENV_dailyupdate/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
idna==3.3
mysql-connector==2.2.9
numpy==1.22.2
pandas==1.4.1
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
six==1.16.0
tqdm==4.62.3
urllib3==1.26.8